+++ src/maicos/lib/math.py
@@ -260,7 +260,7 @@ def correlation_time(
     ----------
     timeseries : numpy.ndarray
         The time series used to calculate the correlation time from.
-    method : {"sokal", "chodera"}
+    method : {``"sokal"``, ``"chodera"``}
         Method to choose summation cutoff :math:`N_\mathrm{cut}`.
     mintime: int
         Minimum possible value for :math:`N_\mathrm{cut}`.
@@ -507,7 +507,10 @@ def center_cluster(ag: mda.AtomGroup, weights: np.ndarray) -> np.ndarray:
 
 
 def symmetrize(
-    m: np.ndarray, axis: None | int | tuple[int] = None, inplace: bool = False
+    m: np.ndarray,
+    axis: None | int | tuple[int] = None,
+    inplace: bool = False,
+    is_odd: bool = False,
 ) -> np.ndarray:
     """Symmeterize an array.
 
@@ -525,6 +528,11 @@ def symmetrize(
          symmetrizing is performed on all of the axes specified in the :obj:`tuple`.
     inplace : bool
         Do symmetrizations inplace. If :obj:`False` a new array is returned.
+    is_odd : bool
+        The parity to use for symmetrization. If :obj:`False` (default), the
+        symmetrization is done with "even" parity, meaning that the output array will be
+        symmetric with respect to the specified axis. If :obj:`True`, the symmetrization
+        is done with "odd" parity, meaning that the output array will be antisymmetric.
 
     Returns
     -------
@@ -546,6 +554,13 @@ def symmetrize(
     array([4.5, 4.5, 4.5, 4.5, 4.5, 4.5, 4.5, 4.5, 4.5, 4.5])
     >>> A
     array([4.5, 4.5, 4.5, 4.5, 4.5, 4.5, 4.5, 4.5, 4.5, 4.5])
+
+    Antisymmetrization can be achieved by setting ``is_odd=True``.
+    >>> A = np.arange(10).astype(float)
+    >>> A
+    array([0., 1., 2., 3., 4., 5., 6., 7., 8., 9.])
+    >>> symmetrize(A, is_odd=True)
+    array([-4.5, -3.5, -2.5, -1.5, -0.5,  0.5,  1.5,  2.5,  3.5,  4.5])
 
     It also works for arrays with more than 1 dimensions in a general dimension.
 
@@ -587,7 +602,7 @@ def symmetrize(
     """
     # The returned array will be of type float
     out = m.copy().astype("float")
-    out += np.flip(m, axis=axis)
+    out += (-1 if is_odd else 1) * np.flip(m, axis=axis)
     out /= 2
 
     if inplace:
@@ -600,7 +615,7 @@ def symmetrize(
     return out
 
 
-def compute_form_factor(q: float, atom_type: str) -> float:
+def compute_form_factor(q: float | np.ndarray, element: str) -> float:
     r"""Calculate the form factor :math:`f(q)`.
 
     :math:`f(q)` is expressed in terms of the scattering vector as
@@ -610,56 +625,66 @@ def compute_form_factor(q: float, atom_type: str) -> float:
 
     The coefficients :math:`a_{1,\dots,4}`, :math:`b_{1,\dots,4}` and :math:`c` are also
     known as Cromer-Mann X-ray scattering factors and are documented in
-    :footcite:t:`princeInternationalTablesCrystallography2004` and taken from
-    https://lampz.tugraz.at/~hadley/ss1/crystaldiffraction/atomicformfactors/formfactors.php.
-    and stored stored in :obj:`maicos.lib.tables.CM_parameters`.
+    :footcite:t:`princeInternationalTablesCrystallography2004` and taken from the `TU
+    Graz
+    <https://lampz.tugraz.at/~hadley/ss1/crystaldiffraction/atomicformfactors/formfactors.php>`_.
+    and stored in :obj:`maicos.lib.tables.CM_parameters`.
 
     Parameters
     ----------
     q : float
         The magnitude of the scattering vector in reciprocal angstroms (1/Å).
-    atom_type : str
-        The type of the atom for which the form factor is calculated. The ``atom_type``
-        is attempted to be converted into an element using
-        :obj:`maicos.lib.tables.atomtypes`. If no suitable element is found, it is taken
-        as is.
+    element : str
+        The element for which the form factor is calculated. Known elements are listed
+        in the :attr:`maicos.lib.tables.elements` set. United-atom models such as
+        ``"CH1"``, ``"CH2"``, ``"CH3"``, ``"CH4"``, ``"NH1"``, ``"NH2"``, and ``"NH3"``
+        are also supported.
+
+        .. note::
+
+            ``element`` is converted to title case to avoid most common issues with
+            MDAnalysis which uses upper case elements by default. For example ``"MG"``
+            will be converted to ``"Mg"``.
 
     Returns
     -------
     float
-        The calculated form factor for the specified atom type and q in units of
+        The calculated form factor for the specified element and q in units of
         electrons.
 
     """
-    if atom_type in tables.atomtypes:
-        element = tables.atomtypes[atom_type]
-    else:
-        element = atom_type
-
     if element == "CH1":
-        form_factor = compute_form_factor(q, "C") + compute_form_factor(q, "H")
-    elif element == "CH2":
-        form_factor = compute_form_factor(q, "C") + 2 * compute_form_factor(q, "H")
-    elif element == "CH3":
-        form_factor = compute_form_factor(q, "C") + 3 * compute_form_factor(q, "H")
-    elif element == "CH4":
-        form_factor = compute_form_factor(q, "C") + 4 * compute_form_factor(q, "H")
-    elif element == "NH1":
-        form_factor = compute_form_factor(q, "N") + compute_form_factor(q, "H")
-    elif element == "NH2":
-        form_factor = compute_form_factor(q, "N") + 2 * compute_form_factor(q, "H")
-    elif element == "NH3":
-        form_factor = compute_form_factor(q, "N") + 3 * compute_form_factor(q, "H")
-    else:
-        form_factor = tables.CM_parameters[element].c
-        # q / (4 * pi) = sin(theta) / lambda
-        q2 = (q / (4 * np.pi)) ** 2
-        for i in range(4):
-            form_factor += tables.CM_parameters[element].a[i] * np.exp(
-                -tables.CM_parameters[element].b[i] * q2
-            )
+        return compute_form_factor(q, "C") + compute_form_factor(q, "H")
+    if element == "CH2":
+        return compute_form_factor(q, "C") + 2 * compute_form_factor(q, "H")
+    if element == "CH3":
+        return compute_form_factor(q, "C") + 3 * compute_form_factor(q, "H")
+    if element == "CH4":
+        return compute_form_factor(q, "C") + 4 * compute_form_factor(q, "H")
+    if element == "NH1":
+        return compute_form_factor(q, "N") + compute_form_factor(q, "H")
+    if element == "NH2":
+        return compute_form_factor(q, "N") + 2 * compute_form_factor(q, "H")
+    if element == "NH3":
+        return compute_form_factor(q, "N") + 3 * compute_form_factor(q, "H")
 
-    return form_factor
+    if element.title() not in tables.CM_parameters:
+        raise ValueError(
+            f"Element '{element}' not found. Known elements are listed in the "
+            "`maicos.lib.tables.elements` set."
+        )
+    # q / (4 * pi) = sin(theta) / lambda
+    q2 = np.asarray((q / (4 * np.pi)) ** 2)
+
+    CM_parameter = tables.CM_parameters[element.title()]
+
+    q2_flat = q2.flatten()
+    form_factor = (
+        np.sum(CM_parameter.a * np.exp(-CM_parameter.b * q2_flat[:, None]), axis=1)
+        + CM_parameter.c
+    )
+
+    return form_factor.reshape(q2.shape)
 
 
 def transform_cylinder(