+++ src/maicos/lib/tables.py
@@ -7,7 +7,8 @@
 # SPDX-License-Identifier: GPL-3.0-or-later
 """The module contains static lookup tables for atom typing etc.
 
-The tables are dictionaries that are indexed by elements.
+The tables are dictionaries that are indexed by elements. All known elements are listed
+in the :attr:`elements` set.
 """
 
 from dataclasses import dataclass
@@ -16,17 +17,6 @@ from pathlib import Path
 import numpy as np
 
 _share_path = Path(__file__).parents[1] / "share"
-
-
-#: Translation of
-#: :py:class:`MDAnalysis.AtomGroup.types <MDAnalysis.core.topologyattrs.Atomtypes>` to
-#: chemical elements.
-atomtypes = {}
-with Path(_share_path / "atomtypes.dat").open() as f:
-    for line in f:
-        if line[0] != "#":
-            elements = line.split()
-            atomtypes[elements[0]] = elements[1]
 
 
 @dataclass
@@ -41,7 +31,7 @@ class CMParameter:
 #: Cromer-Mann X-ray scattering factors computed from numerical
 #: Hartree-Fock wave functions. See https://it.iucr.org/Cb/ch6o1v0001/
 CM_parameters = {}
-with Path(_share_path / "sfactor.dat").open() as f:
+with Path(_share_path / "scatteringfactors.dat").open() as f:
     for line in f:
         if line[0] != "#":
             params = line.split()
@@ -52,3 +42,13 @@ with Path(_share_path / "sfactor.dat").open() as f:
                 b=np.array(params[5:9], dtype=np.double),
                 c=float(params[9]),
             )
+
+#: Set of known elements for Cromer-Mann coefficients.
+elements = set(CM_parameters.keys())
+
+#: Number of electrons for each element
+#: Values are computed from :math:`q=0` limit of Cromer-Mann parameters.
+electron_count = {}
+for element in elements:
+    CM_parameter = CM_parameters[element]
+    electron_count[element] = np.sum(CM_parameter.a) + CM_parameter.c