+++ src/maicos/lib/util.py
@@ -30,9 +30,10 @@ DOC_DICT = dict(
     # DESCRIPTION SECTION
     #####################
     SAVE_METHOD_DESCRIPTION="Save results of analysis to file specified by ``output``.",
-    DENSITY_DESCRIPTION_1=r"""Calculations are carried out for
-``mass`` :math:`(\rm u \cdot Å^{-3})`, ``number`` :math:`(\rm Å^{-3})` or ``charge``
-:math:`(\rm e \cdot Å^{-3})` density profiles """,
+    DENSITY_DESCRIPTION_1=r"""Calculations are carried out for ``mass``
+:math:`(\rm u \cdot Å^{-3})`, ``number`` :math:`(\rm Å^{-3})`, partial ``charge``
+:math:`(\rm e \cdot Å^{-3})` or electron :math:`(\rm e \cdot Å^{-3})` density
+profiles """,
     DENSITY_DESCRIPTION_2="""Cell dimensions are allowed to fluctuate in time.
 
 For grouping with respect to ``molecules``, ``residues`` etc., the corresponding
@@ -103,9 +104,9 @@ self : object
     performed.""",
     WRAP_COMPOUND_PARAMETER="""wrap_compound : str
     The group which will be kept together through the wrap processes. Allowed values
-    are: ``"atoms"``, ``"group"``, ``"residues"``,
-``"segments"``, ``"molecules"``, or ``"fragments"``.""",
-    DENS_PARAMETER="""dens : {``"mass"``, ``"number"``, ``"charge"``}
+    are: ``"atoms"``, ``"group"``, ``"residues"``, ``"segments"``, ``"molecules"``, or
+    ``"fragments"``.""",
+    DENS_PARAMETER="""dens : {``"mass"``, ``"number"``, ``"charge"``, ``"electron"``}
     density type to be calculated.""",
     TEMPERATURE_PARAMETER="""temperature : float
     Reference temperature (K)""",
@@ -113,7 +114,7 @@ self : object
     Width of the bins (in Å).""",
     DIM_PARAMETER="""dim : {0, 1, 2}
     Dimension for binning (``x=0``, ``y=1``, ``z=1``).""",
-    VDIM_PARAMETER="""    vdim : {0, 1, 2}
+    VDIM_PARAMETER="""vdim : {0, 1, 2}
     Dimension for velocity binning (``x=0``, ``y=1``, ``z=1``).""",
     PDIM_PLANAR_PARAMETER="""pdim : {0, 1, 2}
     direction of the projection""",