+++ src/maicos/modules/velocityplanar.py
@@ -34,8 +34,12 @@ class VelocityPlanar(ProfilePlanarBase):
     Parameters
     ----------
     ${PROFILE_PLANAR_CLASS_PARAMETERS}
+    sym_odd : bool,
+        Parity of the profile. If :obj:`False`, the profile will be symmetrized. If
+        :obj:`True`, the profile is antisymmetrized. Only relevant in combination with
+        ``sym``.
     ${VDIM_PARAMETER}
-    $FLUX_PARAMETER}
+    ${FLUX_PARAMETER}
 
     Attributes
     ----------
@@ -52,6 +56,7 @@ class VelocityPlanar(ProfilePlanarBase):
         bin_width: float = 1.0,
         refgroup: mda.AtomGroup | None = None,
         sym: bool = False,
+        sym_odd: bool = False,
         grouping: str = "atoms",
         unwrap: bool = True,
         pack: bool = True,
@@ -79,6 +84,7 @@ class VelocityPlanar(ProfilePlanarBase):
             bin_width=bin_width,
             refgroup=refgroup,
             sym=sym,
+            sym_odd=sym_odd,
             grouping=grouping,
             bin_method=bin_method,
             output=output,