+++ CHANGELOG.rst
@@ -14,6 +14,13 @@ Unreleased
 Philip Loche, Henrik Stooß
 
 - Implement error calculation for observables with multiple samples (#494)
+- Fix wrong equation in ``compute_structure_factor`` documentation (#501)
+- Improve issue templates (#500)
+- Speed up electron density calculation by using a dictionary lookup (#498)
+- Introduce `sym_odd` option to `ProfilePlanarBase` to support vectorial
+  observables (#495)
+- Base Cromer-Mann form factor on ``atomgroup.elements`` (#492)
+- Add electron density weights for density modules (#493)
 - Use gallery view for examples (#491)
 - Create logic for stable and latest documentation versions (#489)
 - Cleanup logos (#487)